+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## \[Next\]
 
+## \[5.6.3\]
+
 - Increase project metadata timeouts.
 - Add cluster cloud logging fiter output.
 - Add Ubuntu 22.04 LTS support.